tests referer
